==> code/00-colors.R <==

colors <- list(
  BCP = list(
    BCP1 = "aquamarine4",
    BCP2 = "aquamarine3",
    BCP3 = "aquamarine"
  ),
  WB = "darkolivegreen4",
  EIRLI = list(
    EIRLI = "darkorchid4",
    DXno = "darkorchid",
    DX0  = "hotpink3",
    DXyes = "firebrick2"
  )
)
